Provides an impl. of exception handling middleware.
For #175
